Also use ModularTenantEvents to ensure that an hash key exists
