+++ src/OrchardCore.Modules/OrchardCore.Media/Processing/MediaTokenOptionsConfiguration.cs
@@ -1,4 +1,3 @@
-using System.Security.Cryptography;
 using Microsoft.Extensions.Options;
 using OrchardCore.Entities;
 using OrchardCore.Settings;
@@ -16,24 +15,10 @@ namespace OrchardCore.Media.Processing
 
         public void Configure(MediaTokenOptions options)
         {
-            var mediaTokenSettings = _siteService.GetSiteSettingsAsync()
+            options.HashKey = _siteService.GetSiteSettingsAsync()
                 .GetAwaiter().GetResult()
-                .As<MediaTokenSettings>();
-
-            if (mediaTokenSettings.HashKey == null)
-            {
-                var siteSettings = _siteService.LoadSiteSettingsAsync().GetAwaiter().GetResult();
-
-                var rng = RandomNumberGenerator.Create();
-
-                mediaTokenSettings.HashKey = new byte[64];
-                rng.GetBytes(mediaTokenSettings.HashKey);
-                siteSettings.Put(mediaTokenSettings);
-
-                _siteService.UpdateSiteSettingsAsync(siteSettings).GetAwaiter().GetResult();
-            }
-
-            options.HashKey = mediaTokenSettings.HashKey;
+                .As<MediaTokenSettings>()
+                .HashKey;
         }
     }
 }
